Module t0rn_collator::chain_spec

Structs

Extensions

The extensions for the ChainSpec.

Constants

PARACHAIN_ID 🔒

REVERT_BYTECODE 🔒

SAFE_XCM_VERSION 🔒

The default XCM version to set in genesis config.

Functions

development_config

fetch_gtwy_init_data 🔒

Fetches gateway initialization data by chain id.

fetch_xdns_record_from_rpc 🔒

Helper function that fetches metadata from live networks and generates a XdnsRecord.

get_account_id_from_adrs

Helper function to derive an account ID from a SS58 address.

get_account_id_from_seed

Helper function to generate an account ID from seed.

get_aura_id_from_adrs

Derive an Aura id from a SS58 address.

get_collator_keys_from_seed

Generate collator keys from seed.

get_public_from_adrs

Helper function to derive a public key from a SS58 address.

get_public_from_seed

Helper function to generate a crypto pair from seed.

initial_gateways 🔒

Lists initialization data for indicated gateways.

is_relaychain 🔒

local_testnet_config

rococo_config

seed_xdns_registry 🔒

Helper function to generate XdnsRecords from RPC.

session_keys

Generate the session keys from individual elements.

standard_side_effects 🔒

testnet_genesis 🔒

Type Definitions

AccountPublic 🔒

ChainSpec

t3rn-pallets chain spec config – END Specialized ChainSpec for the normal parachain runtime.